+++ src/greynoise/api.py
@@ -1,5 +1,6 @@
 """GreyNoise API client."""
 
+import re
 from collections import OrderedDict
 
 import cachetools
@@ -33,7 +34,6 @@ class GreyNoise(object):
     EP_GNQL = "experimental/gnql"
     EP_GNQL_STATS = "experimental/gnql/stats"
     EP_INTERESTING = "interesting/{ip_address}"
-    EP_NOISE_QUICK = "noise/quick/{ip_address}"
     EP_NOISE_MULTI = "noise/multi/quick"
     EP_NOISE_CONTEXT = "noise/context/{ip_address}"
     EP_NOT_IMPLEMENTED = "request/{subcommand}"
@@ -68,6 +68,13 @@ class GreyNoise(object):
     IP_CONTEXT_CACHE = cachetools.TTLCache(maxsize=CACHE_MAX_SIZE, ttl=CACHE_TTL)
 
     IP_QUICK_CHECK_CHUNK_SIZE = 1000
+
+    IPV4_REGEX = re.compile(
+        r"(?:{octet}\.){{3}}{octet}".format(
+            octet=r"(?:(?:25[0-5])|(?:2[0-4]\d)|(?:1?\d?\d))"
+        )
+    )
+    FILTER_TEXT_CHUNK_SIZE = 10000
 
     def __init__(self, api_key=None, timeout=None, use_cache=True):
         if api_key is None or timeout is None:
@@ -130,6 +137,97 @@ class GreyNoise(object):
             raise RequestFailure(response.status_code, body)
 
         return body
+
+    def filter(self, text, noise_only=False):
+        """Filter lines that contain IP addresses from a given text.
+
+        :param text: Text input
+        :type text: str
+        :param noise_only:
+            If set, return only lines that contain IP addresses classified as noise,
+            otherwise, return lines that contain IP addresses not classified as noise.
+        :type noise_only: bool
+        :return: Iterator that yields lines in chunks
+        :rtype: iterable
+
+        """
+        chunks = more_itertools.chunked(text, self.FILTER_TEXT_CHUNK_SIZE)
+        for chunk in chunks:
+            yield self._filter_chunk(chunk, noise_only)
+
+    def _filter_chunk(self, text, noise_only):
+        """Filter chunk of lines that contain IP addresses from a given text.
+
+        :param text: Text input
+        :type text: str
+        :param noise_only:
+            If set, return only lines that contain IP addresses classified as noise,
+            otherwise, return lines that contain IP addresses not classified as noise.
+        :type noise_only: bool
+        :return: Filtered line
+
+        """
+        text_ip_addresses = set()
+        for input_line in text:
+            text_ip_addresses.update(self.IPV4_REGEX.findall(input_line))
+
+        noise_ip_addresses = {
+            result["ip"] for result in self.quick(text_ip_addresses) if result["noise"]
+        }
+
+        def all_ip_addresses_noisy(line):
+            """Select lines that contain IP addresses and all of them are noisy.
+
+            :param line: Line being processed.
+            :type line: str
+            :return: True if line contains IP addresses and all of them are noisy.
+            :rtype: bool
+
+            """
+            line_ip_addresses = self.IPV4_REGEX.findall(line)
+            return line_ip_addresses and all(
+                line_ip_address in noise_ip_addresses
+                for line_ip_address in line_ip_addresses
+            )
+
+        def add_markup(match):
+            """Add markup to surround IP address value with proper tag.
+
+            :param match: IP address match
+            :type match: re.Match
+            :return: IP address with markup
+            :rtype: str
+
+            """
+            ip_address = match.group(0)
+            if ip_address in noise_ip_addresses:
+                tag = "noise"
+            else:
+                tag = "not-noise"
+
+            return "<{tag}>{ip_address}</{tag}>".format(ip_address=ip_address, tag=tag)
+
+        if noise_only:
+            line_matches = all_ip_addresses_noisy
+        else:
+
+            def line_matches(line):
+                """Match all lines that contain either text or non-noisy lines.
+
+                :param line: Line being processed.
+                :type line: str
+                :return: True if line matches as expected.
+                :rtype: bool
+
+                """
+                return not all_ip_addresses_noisy(line)
+
+        filtered_lines = [
+            self.IPV4_REGEX.subn(add_markup, input_line)[0]
+            for input_line in text
+            if line_matches(input_line)
+        ]
+        return "".join(filtered_lines)
 
     def interesting(self, ip_address):
         """Report an IP as "interesting".
@@ -233,19 +331,15 @@ class GreyNoise(object):
             ]
             if api_ip_addresses:
                 api_results = []
-                if len(api_ip_addresses) == 1:
-                    endpoint = self.EP_NOISE_QUICK.format(
-                        ip_address=api_ip_addresses[0]
-                    )
-                    api_results.append(self._request(endpoint))
-                else:
-                    chunks = more_itertools.chunked(
-                        api_ip_addresses, self.IP_QUICK_CHECK_CHUNK_SIZE
-                    )
-                    for chunk in chunks:
-                        api_results.extend(
-                            self._request(self.EP_NOISE_MULTI, json={"ips": chunk})
-                        )
+                chunks = more_itertools.chunked(
+                    api_ip_addresses, self.IP_QUICK_CHECK_CHUNK_SIZE
+                )
+                for chunk in chunks:
+                    api_result = self._request(self.EP_NOISE_MULTI, json={"ips": chunk})
+                    if isinstance(api_result, list):
+                        api_results.extend(api_result)
+                    else:
+                        api_results.append(api_result)
 
                 for api_result in api_results:
                     ip_address = api_result["ip"]
@@ -255,17 +349,15 @@ class GreyNoise(object):
             results = list(ordered_results.values())
         else:
             results = []
-            if len(ip_addresses) == 1:
-                endpoint = self.EP_NOISE_QUICK.format(ip_address=ip_addresses[0])
-                results.append(self._request(endpoint))
-            else:
-                chunks = more_itertools.chunked(
-                    ip_addresses, self.IP_QUICK_CHECK_CHUNK_SIZE
-                )
-                for chunk in chunks:
-                    results.extend(
-                        self._request(self.EP_NOISE_MULTI, json={"ips": chunk})
-                    )
+            chunks = more_itertools.chunked(
+                ip_addresses, self.IP_QUICK_CHECK_CHUNK_SIZE
+            )
+            for chunk in chunks:
+                result = self._request(self.EP_NOISE_MULTI, json={"ips": chunk})
+                if isinstance(result, list):
+                    results.extend(result)
+                else:
+                    results.append(result)
 
         for result in results:
             code = result["code"]